+++ libs/form-component/src/app-components/index.ts
@@ -1,5 +1,5 @@
 export * from './AccordionItem';
 export * from './Button';
 export * from './Card';
-export * from './Date';
+export * from './DisplayDate';
 export * from './Spinner';
